Minor fixes
We should always call try_responding() from the the server itself in
case there were no goals and therefore no callbacks to invoke. I suppose
one alternative to this would be to check in the shared state's
destructor...

Also, use the on_response_ callback as a flag to check if we've already
responded. This should never happen, but just in case...