Move responsibility for responding to dtor
Rather than making the ServerBase explicitly handle CancelRequest
response when there are no relevant goals, we let the
ServerCancelRequestHandleSharedState object handle this implicitly in
its destructor. This will only trigger when the SharedState goes out of
scope and had no cancelled goals, so it's not an issue when there are
cancelled goals.