Make create_server generic over CancelCallback asynchrony
This still has the same problem as for the GoalCallback, where code that
calls these with the result of std::bind() becomes ambiguous. This is
due to the (poor) design of std::bind() returning an object invocable on
any number of arguments. Using std::bind_front() and std::bind_back()
solves this.