Fix synchronous callbacks with std::bind
The returned value of calling std::bind is an object that is callable
with any number of arguments (those which don't correspond to any of the
placeholders given to the std::bind call are evaluated but discarded).
It is therefore ambiguous when trying to cast to either a sync or an
async callback – the final handle argument to an async callback could be
discarded, allowing sync callbacks to look like async ones.

There are workarounds to this. By using std::bind_front and
std::bind_back, you don't get this behaviour since the returned object
is callable only with a fixed number of arguments. These seem to
generally be recommended over std::bind nowadays.

However this doesn't fix the problem for existing code. To avoid
breaking code that passes std::bind results into create_server, this
change adds compile-time logic to determine when a passed callable could
be invoked as a synchronous callback. If so, the callable is cast to the
synchronous callback type (thereby locking in its signature compared to
the std::bind result). Otherwise, it is treated as an asynchronous
callback and similarly cast to the corresponding type. This seems to
allow for easy use of either signature without breaking existing code.