Fix some exception unsafety
Certain callbacks and behaviours are only meant to be executed once,
even in the presence of errors. However, if an exception is thrown
during its execution, the variables holding these callbacks (and shared
states) aren't cleared, meaning they can be erroniously executed again.

This adds better handling to these cases, where we move the callback and
shared state ownership onto the stack rather than inside the handle
objects, ensuring that the callbacks and shared state are properly freed
in the case of an exception being thrown.